Create README.md
This code defines an FTP server address and port, a list of directories to scan, and a list of common filenames to look for. It then loops through each directory, connects to the FTP server and logs in anonymously, changes to the current directory, and loops through each filename to check if it exists. If a file is found, it adds it to a list of found files. Finally, it prints the list of found files or a message indicating that no files were found. You can modify this code to scan for other directories and filenames or to perform other operations on the found files using the various functions provided by the ftplib library.